+++ src/main/resources/local/schema.sql
@@ -4,4 +4,4 @@ create table symbol (id bigint identity primary key, symbol varchar(15) unique, 
 create table dailyquote (id bigint identity primary key,symbol varchar(15), open decimal(12,4), high decimal(12,4), low decimal(12,4), close decimal(12,4), volume bigint, day date, symbol_id bigint, foreign key (symbol_id) references symbol(id));
 create table intradayquote (id bigint identity primary key, symbol varchar(15), open decimal(12,4), high decimal(12,4), low decimal(12,4), close decimal(12,4), volume bigint, localdatetime timestamp, symbol_id bigint, foreign key (symbol_id) references symbol(id));
 create table portfolio (id bigint identity primary key, user_id bigint, name varchar(50), foreign key (user_id) references appuser(id));
-create table portfoliotosymbol (id bigint identity primary key, portfolio_id bigint, symbol_id bigint, weight decimal(12,4), foreign key (portfolio_id) references portfolio(id), foreign key (symbol_id) references symbol(id));
+create table portfoliotosymbol (id bigint identity primary key, portfolio_id bigint, symbol_id bigint, weight bigint, foreign key (portfolio_id) references portfolio(id), foreign key (symbol_id) references symbol(id));
